requirements.txt
mypy
python-binance
pyyaml
Unidecode
